Don't clear catalog commit information in search documents (#586)
This is necessary because the Auxiliary2AzureSearch job could potentially clear all last commit timestamps on all documents which would lead to an alert being fired on missing last commit timestamp in /search/diag.
As a side effect, this also leads to better traceability on the changes made to search documents.
Progress on https://github.com/NuGet/NuGetGallery/issues/6458
