+++ src/NuGet.Services.AzureSearch/Models/ICommittedDocument.cs
@@ -8,13 +8,9 @@ namespace NuGet.Services.AzureSearch
     /// <summary>
     /// A document that has data committed from the catalog.
     /// </summary>
-    public interface ICommittedDocument : IKeyedDocument
+    public interface ICommittedDocument : IUpdatedDocument
     {
-        DateTimeOffset? LastUpdatedDocument { get; set; }
-        string LastDocumentType { get; set; }
-        bool? LastUpdatedFromCatalog { get; set; }
         DateTimeOffset? LastCommitTimestamp { get; set; }
         string LastCommitId { get; set; }
-        void SetLastUpdatedDocumentOnNextRead();
     }
 }
