+++ src/NuGet.Services.AzureSearch/Models/SearchDocument.cs
@@ -75,7 +75,7 @@ namespace NuGet.Services.AzureSearch
         /// parent classes handle this.
         /// </summary>
         [SerializePropertyNamesAsCamelCase]
-        public class UpdateOwners : CommittedDocument, IOwners
+        public class UpdateOwners : UpdatedDocument, IOwners
         {
             public string[] Owners { get; set; }
         }
@@ -86,7 +86,7 @@ namespace NuGet.Services.AzureSearch
         /// <see cref="Full"/> and its parent classes handle this.
         /// </summary>
         [SerializePropertyNamesAsCamelCase]
-        public class UpdateDownloadCount : CommittedDocument, IDownloadCount
+        public class UpdateDownloadCount : UpdatedDocument, IDownloadCount
         {
             public long? TotalDownloadCount { get; set; }
             public double? DownloadScore { get; set; }
@@ -105,7 +105,7 @@ namespace NuGet.Services.AzureSearch
         /// <summary>
         /// Allows index updating code to apply a new list of owners to a document.
         /// </summary>
-        public interface IOwners : ICommittedDocument
+        public interface IOwners : IUpdatedDocument
         {
             string[] Owners { get; set; }
         }
@@ -113,7 +113,7 @@ namespace NuGet.Services.AzureSearch
         /// <summary>
         /// Allows index updating code to apply new download count information to a document.
         /// </summary>
-        public interface IDownloadCount : ICommittedDocument
+        public interface IDownloadCount : IUpdatedDocument
         {
             long? TotalDownloadCount { get; set; }
             double? DownloadScore { get; set; }
