+++ tests/NuGet.Services.AzureSearch.Tests/SearchDocumentBuilderFacts.cs
@@ -168,8 +168,6 @@ namespace NuGet.Services.AzureSearch
       ""lastUpdatedDocument"": ""2018-12-14T09:30:00+00:00"",
       ""lastDocumentType"": ""NuGet.Services.AzureSearch.SearchDocument+UpdateOwners"",
       ""lastUpdatedFromCatalog"": false,
-      ""lastCommitTimestamp"": null,
-      ""lastCommitId"": null,
       ""key"": ""windowsazure_storage-d2luZG93c2F6dXJlLnN0b3JhZ2U1-IncludePrereleaseAndSemVer2""
     }
   ]
@@ -202,8 +200,6 @@ namespace NuGet.Services.AzureSearch
       ""lastUpdatedDocument"": ""2018-12-14T09:30:00+00:00"",
       ""lastDocumentType"": ""NuGet.Services.AzureSearch.SearchDocument+UpdateDownloadCount"",
       ""lastUpdatedFromCatalog"": false,
-      ""lastCommitTimestamp"": null,
-      ""lastCommitId"": null,
       ""key"": ""windowsazure_storage-d2luZG93c2F6dXJlLnN0b3JhZ2U1-IncludePrereleaseAndSemVer2""
     }
   ]
@@ -247,11 +243,11 @@ namespace NuGet.Services.AzureSearch
       ],
       ""isLatestStable"": " + isLatestStable.ToString().ToLowerInvariant() + @",
       ""isLatest"": " + isLatest.ToString().ToLowerInvariant() + @",
+      ""lastCommitTimestamp"": ""2018-12-13T12:30:00+00:00"",
+      ""lastCommitId"": ""6b9b24dd-7aec-48ae-afc1-2a117e3d50d1"",
       ""lastUpdatedDocument"": ""2018-12-14T09:30:00+00:00"",
       ""lastDocumentType"": ""NuGet.Services.AzureSearch.SearchDocument+UpdateVersionList"",
       ""lastUpdatedFromCatalog"": true,
-      ""lastCommitTimestamp"": ""2018-12-13T12:30:00+00:00"",
-      ""lastCommitId"": ""6b9b24dd-7aec-48ae-afc1-2a117e3d50d1"",
       ""key"": ""windowsazure_storage-d2luZG93c2F6dXJlLnN0b3JhZ2U1-IncludePrereleaseAndSemVer2""
     }
   ]
@@ -300,11 +296,11 @@ namespace NuGet.Services.AzureSearch
       ],
       ""isLatestStable"": " + isLatestStable.ToString().ToLowerInvariant() + @",
       ""isLatest"": " + isLatest.ToString().ToLowerInvariant() + @",
+      ""lastCommitTimestamp"": ""2018-12-13T12:30:00+00:00"",
+      ""lastCommitId"": ""6b9b24dd-7aec-48ae-afc1-2a117e3d50d1"",
       ""lastUpdatedDocument"": ""2018-12-14T09:30:00+00:00"",
       ""lastDocumentType"": ""NuGet.Services.AzureSearch.SearchDocument+UpdateVersionListAndOwners"",
       ""lastUpdatedFromCatalog"": true,
-      ""lastCommitTimestamp"": ""2018-12-13T12:30:00+00:00"",
-      ""lastCommitId"": ""6b9b24dd-7aec-48ae-afc1-2a117e3d50d1"",
       ""key"": ""windowsazure_storage-d2luZG93c2F6dXJlLnN0b3JhZ2U1-IncludePrereleaseAndSemVer2""
     }
   ]
@@ -409,11 +405,11 @@ namespace NuGet.Services.AzureSearch
       ],
       ""title"": ""Windows Azure Storage"",
       ""tokenizedPackageId"": ""WindowsAzure.Storage"",
+      ""lastCommitTimestamp"": ""2018-12-13T12:30:00+00:00"",
+      ""lastCommitId"": ""6b9b24dd-7aec-48ae-afc1-2a117e3d50d1"",
       ""lastUpdatedDocument"": ""2018-12-14T09:30:00+00:00"",
       ""lastDocumentType"": ""NuGet.Services.AzureSearch.SearchDocument+UpdateLatest"",
       ""lastUpdatedFromCatalog"": true,
-      ""lastCommitTimestamp"": ""2018-12-13T12:30:00+00:00"",
-      ""lastCommitId"": ""6b9b24dd-7aec-48ae-afc1-2a117e3d50d1"",
       ""key"": ""windowsazure_storage-d2luZG93c2F6dXJlLnN0b3JhZ2U1-" + expected + @"""
     }
   ]
@@ -660,11 +656,11 @@ namespace NuGet.Services.AzureSearch
       ],
       ""title"": ""Windows Azure Storage"",
       ""tokenizedPackageId"": ""WindowsAzure.Storage"",
+      ""lastCommitTimestamp"": null,
+      ""lastCommitId"": null,
       ""lastUpdatedDocument"": ""2018-12-14T09:30:00+00:00"",
       ""lastDocumentType"": ""NuGet.Services.AzureSearch.SearchDocument+Full"",
       ""lastUpdatedFromCatalog"": false,
-      ""lastCommitTimestamp"": null,
-      ""lastCommitId"": null,
       ""key"": ""windowsazure_storage-d2luZG93c2F6dXJlLnN0b3JhZ2U1-" + expected + @"""
     }
   ]
@@ -768,7 +764,7 @@ namespace NuGet.Services.AzureSearch
                 Assert.Empty(allSearchFilters.Except(testedSearchFilters));
             }
 
-            public void SetDocumentLastUpdated(ICommittedDocument document)
+            public void SetDocumentLastUpdated(IUpdatedDocument document)
             {
                 Data.SetDocumentLastUpdated(document, _output);
             }
